Build plugins for both ES 6.2 and 6.3 in same branch.
This handles differences between the 6.2 and 6.3 server API by using a ServiceLoader to load an instance of the ElasticAPI class which handles abstracting the differences in the APIs between versions.

With this change there will be a single mainline branch that can build all plugins. The build output of this package will be:

build/plugins
├── 6.2
│   └── aes-monitoring
│       ├── ...
└── 6.3
    └── aes-monitoring
        ├── ...

We will have another package (maybe named AESMonitoringPlugin_AMI_Env?) with multiple branches (one per ES version) that copies the appropriate version of the aes-monitoring plugin to it's plugins/aes-monitoring directory so that it can be installed into ES.

cr https://code.amazon.com/reviews/CR-3884327