[AESAlerting-160] Add more integration tests for MonitorRunner and enable the ones that were disabled.
This change heavily utilizes the ExecuteMonitor API with the dryrun flag to get the result
of running a monitor synchronously.  As a convenience I have changed how errors
are reported back from this API to allow for reporting all errors that occurred during execution
in a single result. Input errors are now in MonitorRunResult, Painless script errors in
TriggerRunResult and Action error results in ActionRunResult.

cr https://code.amazon.com/reviews/CR-3963137